Port fabric-resource-loader-v0 to 25w14craftmine.
