+++ fabric-resource-loader-v0/src/main/java/net/fabricmc/fabric/mixin/resource/loader/DatapackHandlerMixin.java
@@ -1,0 +1,39 @@
+package net.fabricmc.fabric.mixin.resource.loader;
+
+import net.fabricmc.fabric.impl.resource.loader.BuiltinModResourcePackSource;
+import net.fabricmc.fabric.impl.resource.loader.ModNioResourcePack;
+
+import net.minecraft.class_10958;
+import net.minecraft.resource.ResourcePack;
+import net.minecraft.resource.ResourcePackManager;
+import net.minecraft.resource.ResourcePackProfile;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+import java.util.List;
+
+@Mixin(class_10958.class)
+public class DatapackHandlerMixin {
+	@Redirect(method = "method_68918(Lnet/minecraft/resource/ResourcePackManager;Lnet/minecraft/resource/DataConfiguration;ZZ)Lnet/minecraft/resource/DataConfiguration;", at = @At(value = "INVOKE", target = "Ljava/util/List;contains(Ljava/lang/Object;)Z"))
+	private static boolean onCheckDisabled(List<String> list, Object o, ResourcePackManager resourcePackManager) {
+		String profileId = (String) o;
+		boolean contains = list.contains(profileId);
+
+		if (contains) {
+			return true;
+		}
+
+		ResourcePackProfile profile = resourcePackManager.getProfile(profileId);
+
+		if (profile.getSource() instanceof BuiltinModResourcePackSource) {
+			try (ResourcePack pack = profile.createResourcePack()) {
+				// Prevents automatic load for built-in data packs provided by mods.
+				return pack instanceof ModNioResourcePack modPack && !modPack.getActivationType().isEnabledByDefault();
+			}
+		}
+
+		return false;
+	}
+}
