+++ fabric-resource-loader-v0/src/main/resources/fabric-resource-loader-v0.mixins.json
@@ -6,7 +6,8 @@
     "DatapackCommandMixin",
     "KeyedResourceReloadListenerMixin",
     "LifecycledResourceManagerImplMixin",
-    "MinecraftServerMixin",
+    "DatapackHandlerMixin",
+    "GameInstanceMixin",
     "ResourceMixin",
     "ResourcePackManagerMixin",
     "ResourcePackProfileMixin",
